Part 3 & 4 (Read Desc)
Added functionality to calculate the daily waste emissions (part 3) as well as the total carbon emissions (part 4)

Fixed error in part 5 where there was an undeclared variable being used
Refactored a lot of code.
void functions were converted to double functions so that they return a value, as there isn't any way for me to add up the daily transportation, energy, and waste emissions for the total carbon footprint otherwise.
Added comments to the functions to explain what they do, if the comments suck feel free to change them :P
Commented out the big while loop for the main menu, thinking we might not need it for reasons outlined in the code